separated collectionEntitiesMode and entityListViews components
