+++ ui/src/components/Collection/CollectionEntitiesMode.jsx
@@ -8,94 +8,17 @@ import { withRouter } from 'react-router';
 import queryString from 'query-string';
 
 import { Count, Schema } from 'components/common';
-import EntityListManager from 'components/Entity/EntityListManager';
+import EntityListViews from 'components/Entity/EntityListViews';
 import { queryCollectionEntities } from 'queries';
 import { selectModel } from 'selectors';
 
-import './CollectionEntitiesMode.scss';
-
-const messages = defineMessages({
-  addSchemaPlaceholder: {
-    id: 'collection.addSchema.placeholder',
-    defaultMessage: 'Add new entity type',
-  },
-});
-
 
 class CollectionEntitiesMode extends React.PureComponent {
-  constructor(props) {
-    super(props);
-    this.handleTabChange = this.handleTabChange.bind(this);
-  }
-
-  handleTabChange(type) {
-    const { history, location } = this.props;
-    const parsedHash = queryString.parse(location.hash);
-    parsedHash.type = type;
-
-    history.push({
-      pathname: location.pathname,
-      search: "",
-      hash: queryString.stringify(parsedHash),
-    });
-  }
-
-  renderTable() {
-    const { collection, activeSchema } = this.props;
-    return <EntityListManager
-      collection={collection}
-      schema={activeSchema}
-      onStatusChange={() => {}}
-    />;
-  }
-
   render() {
-    const { collection, activeType, schemaViews, selectableSchemata, intl } = this.props;
-    const showSchemaSelect = collection.writeable && selectableSchemata.length;
-
     return (
-      <Tabs
-        id="CollectionEntitiesModeTabs"
-        className="CollectionEntitiesMode__tabs info-tabs-padding"
-        onChange={this.handleTabChange}
-        selectedTabId={activeType}
-        renderActiveTabPanelOnly
-        vertical
-      >
-        {schemaViews.map(ref => (
-          <Tab
-            id={ref.schema}
-            key={ref.schema}
-            className="CollectionEntitiesMode__tab"
-            title={
-              <>
-                <Schema.Label schema={ref.schema} plural icon />
-                <Count count={ref.count} />
-              </>}
-            panel={this.renderTable()}
-          />
-        ))}
-        {schemaViews.length > 0 && showSchemaSelect && <MenuDivider />}
-        {showSchemaSelect && (
-          <Tab
-            id="new"
-            key="new"
-            disabled
-            className="CollectionEntitiesMode__tab schema-add-tab"
-            title={
-              <Schema.Select
-                onSelect={this.handleTabChange}
-                optionsFilter={schema => selectableSchemata.indexOf(schema.name) !== -1}
-              >
-                <Button
-                  icon="plus"
-                  text={intl.formatMessage(messages.addSchemaPlaceholder)}
-                />
-              </Schema.Select>
-            }
-          />
-        )}
-      </Tabs>
+      <EntityListViews
+        {...this.props}
+      />
     );
   }
 }
