+++ ui/src/components/Entity/EntityListViews.scss
@@ -1,7 +1,7 @@
 @import "app/variables.scss";
 @import "app/mixins.scss";
 
-.CollectionEntitiesMode {
+.EntityListViews {
   &__tabs {
     .bp3-tab-list {
       // margin-right: $aleph-content-padding;
